+++ src-plugins/vanilla-model/entity/impl/MovableImpl.cpp
@@ -54,7 +54,7 @@ frts::PointPtr frts::MovableImpl::getPreviousPathPos() const
     return previous;
 }
 
-unsigned int frts::MovableImpl::getSpeed() const
+double frts::MovableImpl::getSpeed() const
 {
     return this->speed;
 }
@@ -68,7 +68,7 @@ void frts::MovableImpl::setPath(PathPtr path)
     current = nullptr;
 }
 
-void frts::MovableImpl::setSpeed(unsigned int speed)
+void frts::MovableImpl::setSpeed(double speed)
 {
     this->speed = speed;
 }
